+++ back-end/PlanApi/planDatabase.sql
@@ -1,28 +1,28 @@
 CREATE TABLE status (
-    id                      serial PRIMARY KEY,
+    id                      bigserial PRIMARY KEY,
     content                 varchar(50) NOT NULL
 );
 
 CREATE TABLE plan (
-    id                      serial PRIMARY KEY,
-    student_id              integer NOT NULL,
-    speciality_id           smallint NOT NULL,
-    status_id               integer REFERENCES status
+    id                      bigserial PRIMARY KEY,
+    student_id              bigint NOT NULL,
+    speciality_id           bigint NOT NULL,
+    status_id               bigint REFERENCES status
 );
 
 
 CREATE TABLE subject (
-    id                      serial PRIMARY KEY,
-    plan_id                 integer REFERENCES plan,
+    id                      bigserial PRIMARY KEY,
+    plan_id                 bigint REFERENCES plan,
     name                    text NOT NULL,
     ects                    smallint NOT NULL,
-    plan_mapping_row        smallint,
+    plan_mapping_row        bigint,
     is_pwr                  boolean NOT NULL
 );
 
 CREATE TABLE user_home_subject (
-    id                      serial PRIMARY KEY,
-    student_id              integer NOT NULL,
+    id                      bigserial PRIMARY KEY,
+    student_id              bigint NOT NULL,
     name                    text NOT NULL,
     ects                    smallint NOT NULL
 );
